Add 3P Chat Canvas working with A2A and A2UI to the angular sample projects folder.  (#154)
* Safely convert text value to string before rendering

In dynamic-component.ts, NumberValue is handled prior to StringValue, so it's likely for the value to be resolved as a number at runtime.

Ensures that non-string values (like numbers) resolved from data bindings do not break the markdown renderer in the Text component.

* Add 3P Chat Canvas component and its demo to A2UI sample.

* Update a2a-renderer.scss

Use `display: contents` on host for a2a-renderer component to make its host element disappaear and make its child behave as a direct child of the parent.

For addressing the extra DOM comment